Moved TaskScheduler classes
